intro: Move all rendering to separate modules
